requirements.txt
flask
pexels_api
pillow